--- e2e/portal/tests/ChangeRegistrationStatus/DeleteRegistrationWithStatusIncluded.spec.ts
@@ -1,64 +1,0 @@
-import test from '@playwright/test';
-
-import { RegistrationStatusEnum } from '@121-service/src/registration/enum/registration-status.enum';
-import { SeedScript } from '@121-service/src/scripts/enum/seed-script.enum';
-import { seedRegistrationsWithStatus } from '@121-service/test/helpers/registration.helper';
-import {
-  getAccessToken,
-  resetDB,
-} from '@121-service/test/helpers/utility.helper';
-import {
-  programIdPV,
-  registrationPV5,
-} from '@121-service/test/registrations/pagination/pagination-data';
-
-import TableComponent from '@121-e2e/portal/components/TableComponent';
-import BasePage from '@121-e2e/portal/pages/BasePage';
-import LoginPage from '@121-e2e/portal/pages/LoginPage';
-import RegistrationsPage from '@121-e2e/portal/pages/RegistrationsPage';
-
-const toastMessage =
-  'The status of 1 registration(s) is being changed to "Deleted" successfully. The status change can take up to a minute to process.';
-
-// Arrange
-test.beforeEach(async ({ page }) => {
-  await resetDB(SeedScript.nlrcMultiple, __filename);
-  const accessToken = await getAccessToken();
-
-  await seedRegistrationsWithStatus(
-    [registrationPV5],
-    programIdPV,
-    accessToken,
-    RegistrationStatusEnum.included,
-  );
-
-  // Login
-  const loginPage = new LoginPage(page);
-  await page.goto('/');
-  await loginPage.login();
-  // Navigate to program
-  const basePage = new BasePage(page);
-  await basePage.selectProgram('NLRC Direct Digital Aid Program (PV)');
-});
-
-test('[34409] Delete registration with status "Included"', async ({ page }) => {
-  const registrations = new RegistrationsPage(page);
-  const tableComponent = new TableComponent(page);
-  // Act
-  await test.step('Delete registration with status "Included"', async () => {
-    await tableComponent.updateRegistrationStatusWithOptions({
-      registrationName: registrationPV5.fullName,
-      status: 'Delete',
-      sendMessage: false,
-    });
-    await registrations.validateToastMessageAndClose(toastMessage);
-  });
-  // Assert
-  await test.step('Validate registration was deleted succesfully', async () => {
-    await tableComponent.filterColumnByDropDownSelection({
-      columnName: 'Registration Status',
-      selection: 'Included',
-    });
-    await tableComponent.assertEmptyTableState();
-  });
-});
