--- e2e/portal/tests/ChangeRegistrationStatus/MovePasFromIncludedToCompleted.spec.ts
@@ -1,89 +1,0 @@
-import test from '@playwright/test';
-
-import { TransactionStatusEnum } from '@121-service/src/payments/transactions/enums/transaction-status.enum';
-import { RegistrationStatusEnum } from '@121-service/src/registration/enum/registration-status.enum';
-import { SeedScript } from '@121-service/src/scripts/enum/seed-script.enum';
-import {
-  doPayment,
-  waitForPaymentTransactionsToComplete,
-} from '@121-service/test/helpers/program.helper';
-import { seedRegistrationsWithStatus } from '@121-service/test/helpers/registration.helper';
-import {
-  getAccessToken,
-  resetDB,
-} from '@121-service/test/helpers/utility.helper';
-import {
-  programIdPV,
-  registrationPvMaxPayment,
-} from '@121-service/test/registrations/pagination/pagination-data';
-
-import TableComponent from '@121-e2e/portal/components/TableComponent';
-import BasePage from '@121-e2e/portal/pages/BasePage';
-import LoginPage from '@121-e2e/portal/pages/LoginPage';
-import RegistrationsPage from '@121-e2e/portal/pages/RegistrationsPage';
-// Arrange
-test.beforeEach(async ({ page }) => {
-  await resetDB(SeedScript.nlrcMultiple, __filename);
-  const accessToken = await getAccessToken();
-
-  await seedRegistrationsWithStatus(
-    [registrationPvMaxPayment],
-    programIdPV,
-    accessToken,
-    RegistrationStatusEnum.included,
-  );
-
-  // Login
-  const loginPage = new LoginPage(page);
-  await page.goto('/');
-  await loginPage.login();
-  // Navigate to program
-  const basePage = new BasePage(page);
-  await basePage.selectProgram('NLRC Direct Digital Aid Program (PV)');
-});
-test('[31211] Move PA(s) from status "Included" to "Completed"', async ({
-  page,
-}) => {
-  const accessToken = await getAccessToken();
-  const paymentReferenceIds = [registrationPvMaxPayment.referenceId];
-  const registrations = new RegistrationsPage(page);
-  const tableComponent = new TableComponent(page);
-
-  // Act
-  await test.step('Validate the status of the registration', async () => {
-    await registrations.validateStatusOfFirstRegistration({
-      status: 'Included',
-    });
-  });
-
-  await test.step('Change status of registratios to "Completed" with doing a payment', async () => {
-    await doPayment({
-      programId: programIdPV,
-      paymentNr: 1,
-      amount: 100,
-      referenceIds: paymentReferenceIds,
-      accessToken,
-    });
-    // Wait for payment transactions to complete
-    await waitForPaymentTransactionsToComplete({
-      programId: programIdPV,
-      paymentReferenceIds,
-      accessToken,
-      maxWaitTimeMs: 4_000,
-      completeStatusses: Object.values(TransactionStatusEnum),
-    });
-  });
-
-  await test.step('Search for the registration with status "Completed"', async () => {
-    await tableComponent.filterColumnByDropDownSelection({
-      columnName: 'Registration Status',
-      selection: 'Completed',
-    });
-  });
-  // Assert
-  await test.step('Validate the status of the registration', async () => {
-    await registrations.validateStatusOfFirstRegistration({
-      status: 'Completed',
-    });
-  });
-});
