+++ e2e/portal/tests/ChangeRegistrationStatus/MoveRegistrationsFromNewToValidated.spec.ts
@@ -1,12 +1,8 @@
 import test from '@playwright/test';
 
-import { RegistrationStatusEnum } from '@121-service/src/registration/enum/registration-status.enum';
 import { SeedScript } from '@121-service/src/scripts/enum/seed-script.enum';
-import { seedRegistrationsWithStatus } from '@121-service/test/helpers/registration.helper';
-import {
-  getAccessToken,
-  resetDB,
-} from '@121-service/test/helpers/utility.helper';
+import { seedRegistrations } from '@121-service/test/helpers/registration.helper';
+import { resetDB } from '@121-service/test/helpers/utility.helper';
 import {
   programIdPV,
   registrationPV5,
@@ -18,19 +14,12 @@ import LoginPage from '@121-e2e/portal/pages/LoginPage';
 import RegistrationsPage from '@121-e2e/portal/pages/RegistrationsPage';
 
 const toastMessage =
-  'The status of 1 registration(s) is being changed to "Deleted" successfully. The status change can take up to a minute to process.';
-
+  'The status of 1 registration(s) is being changed to "Validated" successfully. The status change can take up to a minute to process.';
 // Arrange
 test.beforeEach(async ({ page }) => {
   await resetDB(SeedScript.nlrcMultiple, __filename);
-  const accessToken = await getAccessToken();
 
-  await seedRegistrationsWithStatus(
-    [registrationPV5],
-    programIdPV,
-    accessToken,
-    RegistrationStatusEnum.new,
-  );
+  await seedRegistrations([registrationPV5], programIdPV);
 
   // Login
   const loginPage = new LoginPage(page);
@@ -41,24 +30,31 @@ test.beforeEach(async ({ page }) => {
   await basePage.selectProgram('NLRC Direct Digital Aid Program (PV)');
 });
 
-test('[34408] Delete registration with status "New"', async ({ page }) => {
+test('[31206] Move PA(s) from status "New" to "Validated"', async ({
+  page,
+}) => {
   const registrations = new RegistrationsPage(page);
   const tableComponent = new TableComponent(page);
   // Act
-  await test.step('Delete registration with status "New"', async () => {
+  await test.step('Change status of first selected registration to "Validated"', async () => {
     await tableComponent.updateRegistrationStatusWithOptions({
       registrationName: registrationPV5.fullName,
-      status: 'Delete',
+      status: 'Validate',
       sendMessage: false,
     });
     await registrations.validateToastMessageAndClose(toastMessage);
   });
-  // Assert
-  await test.step('Validate registration was deleted succesfully', async () => {
+
+  await test.step('Search for the registration with status "Validated"', async () => {
     await tableComponent.filterColumnByDropDownSelection({
       columnName: 'Registration Status',
-      selection: 'New',
+      selection: 'Validated',
     });
-    await tableComponent.assertEmptyTableState();
+  });
+  // Assert
+  await test.step('Validate the status of the registration', async () => {
+    await registrations.validateStatusOfFirstRegistration({
+      status: 'Validated',
+    });
   });
 });
