+++ e2e/portal/tests/__screenshots__/ChangeRegistrationStatus/ChangeRegistrationStatusWithMessages.spec.ts/Change-status-of-registration-with-different-s-69c84-status-of-registration-with-templated-message-1.txt
@@ -1,0 +1,11 @@
+Dit is een bericht van het Rode Kruis.
+
+Bedankt voor je inschrijving. Je ontvangt vanaf nu elke dinsdag een Albert Heijn waardebon via WhatsApp. Je ontvangt de waardebonnen zo lang je op de lijst staat van .
+
+Het Rode Kruis kan je ook informatie geven over bijvoorbeeld medische hulp, voedsel of veiligheid. Kijk  op onze website:
+
+https://helpfulinformation.redcross.nl/
+
+of stel je vraag via WhatsApp:
+
+https://wa.me/3197010286964
