+++ e2e/portal/tests/__screenshots__/ChangeRegistrationStatus/ChangeRegistrationStatusWithMessages.spec.ts/Change-status-of-registration-with-different-s-72326-ge-status-of-registration-with-custom-message-1.txt
@@ -1,0 +1,1 @@
+Test custom message to change the status of registration
